+++ persistent-test/src/PersistentTest.hs
@@ -126,7 +126,7 @@ share [mkPersist persistSettings,  mkMigrate "testMigrate", mkDeleteCascade pers
     email Text
     attr Text
     extra Text
-    age Int      
+    age Int
     UniqueUpsert email
     deriving Eq Show
 
@@ -492,6 +492,41 @@ specs = describe "persistent" $ do
       pBlue30 <- updateGet key25 [PersonAge +=. 2]
       pBlue30 @== Person "Updated" 30 Nothing
 
+  describe "putMany" $ do
+    it "adds new rows when no conflicts" $ db $ do
+        let mkUpsert e = Upsert e "new" "" 1
+        let keys = ["putMany1","putMany2","putMany3"]
+        let vals = map mkUpsert keys
+        _ <- putMany vals
+        Just (Entity _ v1) <- getBy $ UniqueUpsert "putMany1"
+        Just (Entity _ v2) <- getBy $ UniqueUpsert "putMany2"
+        Just (Entity _ v3) <- getBy $ UniqueUpsert "putMany3"
+        [v1,v2,v3] @== vals
+        deleteBy $ UniqueUpsert "putMany1"
+        deleteBy $ UniqueUpsert "putMany2"
+        deleteBy $ UniqueUpsert "putMany3"
+    it "handles conflicts by replacing old keys with new records" $ db $ do
+        let mkUpsert1 e = Upsert e "new" "" 1
+        let mkUpsert2 e = Upsert e "new" "" 2
+        let vals = map mkUpsert2 ["putMany4", "putMany5", "putMany6", "putMany7"]
+        Entity k1 _ <- insertEntity $ mkUpsert1 "putMany4"
+        Entity k2 _ <- insertEntity $ mkUpsert1 "putMany5"
+        _ <- putMany $ [mkUpsert1 "putMany4"] ++ vals
+        Just e1 <- getBy $ UniqueUpsert "putMany4"
+        Just e2 <- getBy $ UniqueUpsert "putMany5"
+        Just e3@(Entity k3 _) <- getBy $ UniqueUpsert "putMany6"
+        Just e4@(Entity k4 _) <- getBy $ UniqueUpsert "putMany7"
+
+        [e1,e2,e3,e4] @== [(Entity k1 $ mkUpsert2 "putMany4")
+                          ,(Entity k2 $ mkUpsert2 "putMany5")
+                          ,(Entity k3 $ mkUpsert2 "putMany6")
+                          ,(Entity k4 $ mkUpsert2 "putMany7")
+                          ]
+        deleteBy $ UniqueUpsert "putMany4"
+        deleteBy $ UniqueUpsert "putMany5"
+        deleteBy $ UniqueUpsert "putMany6"
+        deleteBy $ UniqueUpsert "putMany7"
+
   describe "upsert" $ do
     it "adds a new row with no updates" $ db $ do
         Entity _ u <- upsert (Upsert "a" "new" "" 2) [UpsertAttr =. "update"]
@@ -502,12 +537,12 @@ specs = describe "persistent" $ do
 #ifdef WITH_MONGODB
         initial <- insertEntity (Upsert "foo" "initial" "" 2)
         update' <- upsert (Upsert "foo" "update" "" 3) []
-        update' @== initial      
+        update' @== initial
 #else
         initial <- insertEntity (Upsert "a" "initial" "" 1)
         update' <- upsert (Upsert "a" "update" "" 2) []
         update' @== initial
-#endif                        
+#endif
     it "updates an existing row - assignment" $ db $ do
 #ifdef WITH_MONGODB
         initial <- insertEntity (Upsert "cow" "initial" "extra" 1)
@@ -750,7 +785,7 @@ specs = describe "persistent" $ do
 
   it "insertRecord" $ db $ do
       let record = Person "name" 1 Nothing
-      record' <- insertRecord record 
+      record' <- insertRecord record
       record' @== record
 
   it "getEntity" $ db $ do