+++ persistent/Database/Persist/Sql/Orphan/PersistUnique.hs
@@ -2,6 +2,7 @@
 {-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE FlexibleContexts #-}
 {-# OPTIONS_GHC -fno-warn-orphans  #-}
+{-# OPTIONS_GHC -fno-warn-unused-imports #-}
 
 module Database.Persist.Sql.Orphan.PersistUnique
   ()
@@ -11,14 +12,17 @@ import Control.Exception (throwIO)
 import Control.Monad.IO.Class (liftIO, MonadIO)
 import Control.Monad.Trans.Reader (ReaderT)
 import Database.Persist
+import Database.Persist.Class.PersistUnique (defaultPutMany, persistUniqueKeyValues)
 import Database.Persist.Sql.Types
 import Database.Persist.Sql.Raw
 import Database.Persist.Sql.Orphan.PersistStore (withRawQuery)
-import Database.Persist.Sql.Util (dbColumns, parseEntityValues)
+import Database.Persist.Sql.Util (dbColumns, parseEntityValues, updatePersistValue, mkUpdateText')
 import qualified Data.Text as T
-import Data.Monoid (mappend, (<>))
+import Data.Monoid (mappend)
 import qualified Data.Conduit.List as CL
 import Control.Monad.Trans.Reader (ask, withReaderT)
+import Data.List (nubBy)
+import Data.Function (on)
 
 defaultUpsert
     :: (MonadIO m
@@ -30,34 +34,20 @@ defaultUpsert record updates = do
     uniqueKey <- onlyUnique record
     upsertBy uniqueKey record updates
 
-escape :: DBName -> T.Text
-escape (DBName s) = T.pack $ '"' : escapeQuote (T.unpack s) ++ "\""
-  where
-    escapeQuote "" = ""
-    escapeQuote ('"':xs) = "\"\"" ++ escapeQuote xs
-    escapeQuote (x:xs) = x : escapeQuote xs
-
 instance PersistUniqueWrite SqlBackend where
     upsert record updates = do
       conn <- ask
+      let escape = connEscapeName conn
+      let refCol n = T.concat [escape (entityDB t), ".", n]
+      let mkUpdateText = mkUpdateText' escape refCol
       uniqueKey <- onlyUnique record
       case connUpsertSql conn of
         Just upsertSql -> case updates of
                             [] -> defaultUpsert record updates
                             _:_ -> do
-                                let upds = T.intercalate "," $ map (go' . go) updates
+                                let upds = T.intercalate "," $ map mkUpdateText updates
                                     sql = upsertSql t upds
                                     vals = (map toPersistValue $ toPersistFields record) ++ (map updatePersistValue updates) ++ (unqs uniqueKey)
-                                           
-                                    go'' n Assign = n <> "=?"
-                                    go'' n Add = T.concat [n, "=", escape (entityDB t) <> ".", n, "+?"]
-                                    go'' n Subtract = T.concat [n, "=", escape (entityDB t) <> ".", n, "-?"]
-                                    go'' n Multiply = T.concat [n, "=", escape (entityDB t) <> ".", n, "*?"]
-                                    go'' n Divide = T.concat [n, "=", escape (entityDB t) <> ".", n, "/?"]
-                                    go'' _ (BackendSpecificUpdate up) = error $ T.unpack $ "BackendSpecificUpdate" `Data.Monoid.mappend` up `mappend` "not supported"
-                                              
-                                    go' (x, pu) = go'' (connEscapeName conn x) pu
-                                    go x = (fieldDB $ updateFieldDef x, updateUpdate x)
 
                                 x <- rawSql sql vals
                                 return $ head x
@@ -82,8 +72,21 @@ instance PersistUniqueWrite SqlBackend where
                 , " WHERE "
                 , T.intercalate " AND " $ map (go' conn) $ go uniq]
 
+    putMany [] = return ()
+    putMany rsD = do
+        conn <- ask
+        let rs = nubBy ((==) `on` persistUniqueKeyValues) (reverse rsD)
+        let ent = entityDef rs
+        let nr  = length rs
+        let toVals r = (map toPersistValue $ toPersistFields r)
+        case connPutManySql conn of
+            (Just mkSql) -> rawExecute (mkSql ent nr) (concat (map toVals rs))
+            Nothing -> defaultPutMany rs
+
 instance PersistUniqueWrite SqlWriteBackend where
     deleteBy uniq = withReaderT persistBackend $ deleteBy uniq
+    upsert rs us = withReaderT persistBackend $ upsert rs us
+    putMany rs = withReaderT persistBackend $ putMany rs
 
 instance PersistUniqueRead SqlBackend where
     getBy uniq = do
@@ -122,15 +125,3 @@ instance PersistUniqueRead SqlWriteBackend where
 
 dummyFromUnique :: Unique v -> Maybe v
 dummyFromUnique _ = Nothing
-
-updateFieldDef
-    :: PersistEntity v
-    => Update v -> FieldDef
-updateFieldDef (Update f _ _) = persistFieldDef f
-updateFieldDef (BackendUpdate{}) =
-    error "updateFieldDef did not expect BackendUpdate"
-
-updatePersistValue :: Update v -> PersistValue
-updatePersistValue (Update _ v _) = toPersistValue v
-updatePersistValue (BackendUpdate{}) =
-    error "updatePersistValue did not expect BackendUpdate"