+++ persistent/Database/Persist/Sql/Util.hs
@@ -9,18 +9,26 @@ module Database.Persist.Sql.Util (
   , dbIdColumns
   , dbIdColumnsEsc
   , dbColumns
+  , updateFieldDef
+  , updatePersistValue
+  , mkUpdateText
+  , mkUpdateText'
+  , commaSeparated
+  , parenWrapped
 ) where
 
 import Data.Maybe (isJust)
 import Data.Monoid ((<>))
+import qualified Data.Text as T
 import Data.Text (Text, pack)
 import Database.Persist (
     Entity(Entity), EntityDef, EntityField, HaskellName(HaskellName)
   , PersistEntity, PersistValue
   , keyFromValues, fromPersistValues, fieldDB, entityId, entityPrimary
   , entityFields, entityKeyFields, fieldHaskell, compositeFields, persistFieldDef
-  , keyAndEntityFields
-  , DBName)
+  , keyAndEntityFields, toPersistValue, DBName, Update(..), PersistUpdate(..)
+  , FieldDef
+  )
 import Database.Persist.Sql.Types (Sql, SqlBackend, connEscapeName)
 
 entityColumnNames :: EntityDef -> SqlBackend -> [Sql]
@@ -85,3 +93,35 @@ parseEntityValues t vals =
 
 isIdField :: PersistEntity record => EntityField record typ -> Bool
 isIdField f = fieldHaskell (persistFieldDef f) == HaskellName "Id"
+
+-- | Gets the 'FieldDef' for an 'Update'.
+updateFieldDef :: PersistEntity v => Update v -> FieldDef
+updateFieldDef (Update f _ _) = persistFieldDef f
+updateFieldDef BackendUpdate {} = error "updateFieldDef: did not expect BackendUpdate"
+
+updatePersistValue :: Update v -> PersistValue
+updatePersistValue (Update _ v _) = toPersistValue v
+updatePersistValue (BackendUpdate{}) =
+    error "updatePersistValue: did not expect BackendUpdate"
+
+commaSeparated :: [Text] -> Text
+commaSeparated = T.intercalate ", "
+
+mkUpdateText :: PersistEntity record => SqlBackend -> Update record -> Text
+mkUpdateText conn = mkUpdateText' (connEscapeName conn) id
+
+mkUpdateText' :: PersistEntity record => (DBName -> Text) -> (Text -> Text) -> Update record -> Text
+mkUpdateText' escapeName refColumn x =
+  case updateUpdate x of
+    Assign -> n <> "=?"
+    Add -> T.concat [n, "=", refColumn n, "+?"]
+    Subtract -> T.concat [n, "=", refColumn n, "-?"]
+    Multiply -> T.concat [n, "=", refColumn n, "*?"]
+    Divide -> T.concat [n, "=", refColumn n, "/?"]
+    BackendSpecificUpdate up ->
+      error . T.unpack $ "mkUpdateText: BackendSpecificUpdate " <> up <> " not supported"
+  where
+    n = escapeName . fieldDB . updateFieldDef $ x
+
+parenWrapped :: Text -> Text
+parenWrapped t = T.concat ["(", t, ")"]