+++ persistent/ChangeLog.md
@@ -1,15 +1,25 @@
+## 2.8.1
+
+* DRY-ed up and exposed several util functions in `Database.Persist.Sql.Util`.
+	* Upstream-ed `updatePersistValue`, `mkUpdateText`, and `commaSeparated` from `Database.Persist.MySQL`.
+	* De-duplicated `updatePersistValue` from various `Database.Persist.Sql.Orphan.*` modules.
+* Batching enhancements to reduce db round-trips.
+	* Added `getMany` and `repsertMany` for batched `get` and `repsert`.
+	* Added `putMany` with a default/slow implementation. SqlBackend's that support native UPSERT should override this for batching enhancements.
+	* Updated `insertEntityMany` to replace slow looped usage with batched execution.
+
 ## 2.8.0
 
 * Switch from `MonadBaseControl` to `MonadUnliftIO`
 * Reapplies [#723](https://github.com/yesodweb/persistent/pull/723), which was reverted in version 2.7.3.
 
 ## 2.7.3.1
- 
+
  * Improve error messages when failing to parse database results into Persistent records. [#741](https://github.com/yesodweb/persistent/pull/741)
  * A handful of `fromPersistField` implementations called `error` instead of returning a `Left Text`. All of the implementations were changed to return `Left`. [#741](https://github.com/yesodweb/persistent/pull/741)
  * Improve error message when a SQL insert fails with a custom primary key [#757](https://github.com/yesodweb/persistent/pull/757)
 
-## 2.7.3 
+## 2.7.3
 
 * Reverts [#723](https://github.com/yesodweb/persistent/pull/723), which generalized functions using the `BackendCompatible` class. These changes were an accidental breaking change.
 * Recommend the `PersistDbSpecific` docs if someone gets an error about converting from `PersistDbSpecific`